New silver components: Repair, Smite, and Thunder
Repair:
Repairs all currently equipped items at 5 mana/durability. Meant to use
purified vinteum dust for
reagent, but it looks like it's currently broken

Smite:
Does a small amount of holy damage to its target. If the target is
either in creative or is a boss, the damage is greatly amplified, but if
it's a boss, then the user takes the same amount of damage (usually
killing him/herself)

Thunder:
Summons a lightning bolt at the target location.

Note: Holy damage now bypasses all forms of resistance, and affects
creative players